requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-huggingface
langchain-text-splitters
PyPDF2
faiss-cpu
python-dotenv